requirements.txt
pytest-shutil
firebase-admin
scikit-learn
google-generativeai==0.7.2
tavily-python==0.5.0
openai==1.44.1
selenium==4.27.1
beautifulsoup4==4.12.2
requests==2.32.3
python-dotenv==1.0.1
autogen==0.3.0
datatypes==0.19.0
Flask==3.0.3
Flask-Cors==5.0.0
pandas==2.2.3
numpy==1.26.4
firebase-admin==6.6.0